emitter

+++ integration/Context/Request/LocationRequest.cs
@@ -3,8 +3,8 @@
 public class LocationRequest
 {
     public int? idAsuPro { get; set; }
-    public double longitude { get; set; }
-    public double latitude { get; set; }
+    public decimal longitude { get; set; }
+    public decimal latitude { get; set; }
     public string? status { get; set; }
     public string? address { get; set; }
 }
